Now with a data scriptable object specifically for bag ids & Controller hider script having a reference to the data  & ZuidOoost prefab, when spawning hiding the buildings immediately
