+++ Assets/BagIdHider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c12702bc992978b43b2e53562d173b5c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
