+++ Assets/Scriptables/Data/BagIdsHiddenAmsterdamZuidOost.asset
@@ -1,0 +1,41 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 725deca56fed1b24da6faceaf7e8828a, type: 3}
+  m_Name: BagIdsHiddenAmsterdamZuidOost
+  m_EditorClassIdentifier: 
+  bagIds:
+  - 0363100012254297
+  - 0363100100091868
+  - 0363100012122535
+  - 0363100012249777
+  - 0363100012199405
+  - 0363100012571047
+  - 0363100012254334
+  - 0363100012570739
+  - 0363100012155182
+  - 0363100012079990
+  - 0363100100092000
+  - 0363100012571663
+  - 0363100012076532
+  - 0363100012101986
+  - 0363100012063048
+  - 0363100012570880
+  - 0363100012122538
+  - 0363100012253591
+  - 0363100012139825
+  - 0363100012114502
+  - 0363100012077142
+  - 0363100012205758
+  - 0363100012254286
+  - 0363100012123051
+  - 0363100012124178
+  - 0363100012139268
