+++ Assets/Scriptables/Data/BagIdsHiddenAmsterdamZuidOost.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9e00c0494ae876841a666d0dc3db3a6d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
