+++ Assets/_Functionalities/ObjectInformation/Scripts/HiddenBagIds.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Netherlands3D.Twin
+{
+    [CreateAssetMenu(menuName = "Netherlands3D/Data/BagIdsHiddenData", fileName = "BagIdsHidden", order = 0)]
+    public class HiddenBagIds : ScriptableObject
+    {         
+        [SerializeField] public List<string> bagIds = new List<string>();
+    }
+}
